move more tests into their own classes More tests executed as some of them were previously disabled in headless mode.
